Move auth services etc into common/auth
