+++ app/common/auth/authentication-interceptor.config.js
