+++ app/common/auth/password-reset-confirm.controller.js
