+++ app/common/auth/register.controller.js
