+++ app/common/auth/registration.service.js
