+++ app/common/common-routes.js
@@ -7,28 +7,28 @@ module.exports = ['$routeProvider', '$locationProvider', function ($routeProvide
             redirectTo: '/views/map'
         })
         .when('/login', {
-            controller: require('./controllers/login-controller.js'),
+            controller: require('./controllers/auth/login.controller.js'),
             template: ''
         })
         .when('/forbidden', {
-            controller: require('./controllers/forbidden-controller.js'),
-            templateUrl: 'templates/forbidden.html'
+            controller: require('./auth/forbidden.controller.js'),
+            templateUrl: 'templates/auth/forbidden.html'
         })
         .when('/register', {
-            controller: require('./controllers/register-controller.js'),
-            templateUrl: 'templates/register.html'
+            controller: require('./auth/register.controller.js'),
+            templateUrl: 'templates/auth/register.html'
         })
         .when('/forgotpassword', {
-            controller: require('./controllers/password-reset-controller.js'),
-            templateUrl: 'templates/password-reset.html'
+            controller: require('./auth/password-reset.controller.js'),
+            templateUrl: 'templates/auth/password-reset.html'
         })
         .when('/forgotpassword/confirm', {
-            controller: require('./controllers/password-reset-confirm-controller.js'),
-            templateUrl: 'templates/password-reset-confirm.html'
+            controller: require('./auth/password-reset-confirm.controller.js'),
+            templateUrl: 'templates/auth/password-reset-confirm.html'
         })
         .when('/forgotpassword/confirm/:token*', {
-            controller: require('./controllers/password-reset-confirm-controller.js'),
-            templateUrl: 'templates/password-reset-confirm.html'
+            controller: require('./auth/password-reset-confirm.controller.js'),
+            templateUrl: 'templates/auth/password-reset-confirm.html'
         })
         ;
 }];
