+++ server/www/templates/auth/register.html
